PR 피드백 반영 파일 구조 최적화
