+++ package.json
@@ -13,6 +13,7 @@
     "@tailwindcss/postcss": "^4.1.18",
     "@tailwindcss/vite": "^4.1.18",
     "axios": "^1.13.2",
+    "lucide-react": "^0.563.0",
     "msw": "^2.12.7",
     "prettier": "^3.7.4",
     "react": "^19.2.0",
